misc/HtmlTestSuite/results/test-0139.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Tests parsing of the initial-letter property</title>
<link href="https://www.google.com/" rel="author" title="Google LLC"/>
<link href="https://https://w3c.github.io/csswg-drafts/css-inline-3/#sizing-drop-initials" rel="help"/>
<meta content="initial-letter supports the full grammar 'normal | &lt;number [1,∞]&gt; &lt;integer [1,∞]&gt; | &lt;number [1,∞]&gt; &amp;&amp; [ drop | raise ]?'." name="assert"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/css/support/computed-testcommon.js"/>

</head><body><div id="target"/>
<div id="scratch"/>

<script>
test_computed_value('initial-letter', 'normal');
test_computed_value('initial-letter', '1.23');
test_computed_value('initial-letter', '1.23 456');
test_computed_value('initial-letter', '1.23 calc(45.6)', '1.23 46');
test_computed_value('initial-letter', '1.23 drop');
test_computed_value('initial-letter', '1.23 raise');
test_computed_value('initial-letter', 'drop 1.23', '1.23 drop');
test_computed_value('initial-letter', 'raise 1.23', '1.23 raise');
</script>
</body></html>
